% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_to_pid.R
\name{check_version}
\alias{check_version}
\title{Check pid version}
\usage{
check_version(pid)
}
\arguments{
\item{pid}{(character) The pid to a data, metadata, or resource map object on a DataONE member node}
}
\description{
This function takes a pid and checks to see if it has been obsoleted.
}
\examples{
\dontrun{
   pid <- "https://pasta.lternet.edu/package/data/eml/edi/195/2/51abf1c7a36a33a2a8bb05ccbf8c81c6"
   pid <- "doi:10.18739/A2HF7Z"
   
   check_version(pid)
}

}
