% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get tabular metadata}
\usage{
get_metadata(meta_pids, parallelize = FALSE, full = FALSE)
}
\arguments{
\item{meta_pids}{(character vector) Metadata PIDs}

\item{parallelize}{(logical) If TRUE, uses the parallel package to parallelize the proess. Default = FALSE.}

\item{full}{(logical) Whether to extract the full metadata. See \code{tidy_eml}. Default = FALSE.}
}
\description{
Gets tabular metadata from a list of metadata PIDs.
}
\examples{
\dontrun{
metadata <- get_metadata(c("doi:10.18739/A2K86G", "doi:10.18739/A2KG4S"), parallelize = TRUE)
}
}
