% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_object.R
\name{get_object}
\alias{get_object}
\title{Get a DataONE object}
\usage{
get_object(data_pid, as = "parsed", ...)
}
\arguments{
\item{data_pid}{(character) The data or metadata object PID}

\item{as}{desired type of output: raw, text or parsed. content attempts to automatically figure out which one is most appropriate, based on the content-type. (based on \code{httr::content()})}

\item{...}{pass arguments to read.csv}
}
\description{
This function download a DataONE data or metadata object into the R environment
}
\examples{
\dontrun{
   data <- get_object("urn:uuid:a81f49db-5841-4095-aee2-b0cad7a35cc0")
   meta <- get_object("doi:10.18739/A2PC2T79B")
}

}
