% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pids.R
\name{get_pids}
\alias{get_pids}
\title{Get pids}
\usage{
get_pids(pkg_doi, data_url = NULL)
}
\arguments{
\item{pkg_doi}{(character) The package DOI, or the metadata object PID on a DataONE member node #TODO: figure out terminology}

\item{data_url}{(character) Optional. The URL to the data object on a DataONE member node. If the data_url is passed, only the specified data object PIDs will be returned}
}
\description{
This function takes a package DOI and optionally, a data URL and returns the associated pids. Alternative to \code{url_to_pid()}, based on \code{arcticdatautils::get_package()}.
}
\examples{
\dontrun{
#get all pids in the data package
get_pids("https://doi.org/10.6073/pasta/cd7ec5009430eb78f464a66f6794579f")
#return only pids related to the specified data object
get_pids("doi:10.18739/A2PC2T79B", 
         "https://cn.dataone.org/cn/v2/resolve/urn:uuid#' :a81f49db-5841-4095-aee2-b0cad7a35cc0")
get_pids("https://doi.org/10.6073/pasta/cd7ec5009430eb78f464a66f6794579f", 
         "https://portal.lternet.edu/nis/dataviewer?packageid=knb-lter-luq.107.9996700&entityid=6c2bcaf6cdd5b365a51b1d3318f93eef")
}

}
