% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_attributes.R
\name{match_attributes}
\alias{match_attributes}
\title{Match attributes to data}
\usage{
match_attributes(data, attributes)
}
\arguments{
\item{data}{(data.frame) Data frame to match with attribute metadata}

\item{attributes}{(data.frame) Attribute metadata}
}
\description{
Match attribute metadata to data. Returns a dplyr expression that can be proofed and pasted into a script.
}
