% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_eml.R
\name{tidy_eml}
\alias{tidy_eml}
\title{Get tidy (tabular) metadata}
\usage{
tidy_eml(eml, full = FALSE)
}
\arguments{
\item{eml}{An EML class object, the path to an EML (.xml) metadata file, or a raw EML object}

\item{full}{(logical) Returns the most commonly used metadata fields by default. 
If \code{full = TRUE} is specified, the full set of metadata fields are returned.}
}
\description{
This function takes a path to an EML (.xml) metadata file and returns a data frame.
}
\examples{
\dontrun{
   eml <- system.file("example-eml.xml", package = "arcticdatautils")
   tidy_eml(eml)
   tidy_eml(eml, full = TRUE)
}

}
