% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_to_pid.R
\name{url_to_pid}
\alias{url_to_pid}
\title{Get pid from data URL}
\usage{
url_to_pid(url)
}
\arguments{
\item{url}{(character) The URL to the data object on a DataONE member node}
}
\description{
This function takes a data URL and returns the data pid, as well as the associated metadata and resource map pids.
}
\examples{
\dontrun{
   url <- "https://pasta.lternet.edu/package/data/eml/edi/195/2/51abf1c7a36a33a2a8bb05ccbf8c81c6"
   url <- "https://arcticdata.io/metacat/d1/mn/v2/object/urn\%3Auuid\%3Aa81f49db-5841-4095-aee2-b0cad7a35cc0"
   
   url_to_pid(url)
}

}
