% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_object.R
\name{get_object}
\alias{get_object}
\title{Get a DataONE data object}
\usage{
get_object(data_pid, as = "parsed")
}
\arguments{
\item{data_pid}{(character) The data object PID}

\item{as}{desired type of output: raw, text or parsed. content attempts to automatically figure out which one is most appropriate, based on the content-type. (based on \code{httr::content()})}
}
\description{
This function download a DataONE data object into the R environment
}
\examples{
\dontrun{
   get_object("urn:uuid:a81f49db-5841-4095-aee2-b0cad7a35cc0")
}

}
